Rename old apollo resolvers
